# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Content Workflow API
  description: |
    REST API for AI-powered content creation, translation, and review workflows.
    
    This API provides comprehensive endpoints for managing marketing campaigns,
    content pieces, AI generation, translations, and review processes.
    
    **Features:**
    - Campaign and content management
    - Multi-model AI content generation (OpenAI, Anthropic)
    - Automated translation services
    - Content analysis and sentiment detection
    - Review and approval workflows
    - Real-time monitoring and health checks
    
  version: 1.0.0
  contact:
    name: AI Content Workflow Team
    url: https://github.com/your-org/ai-content-workflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
paths:
...
  /content:
    get:
      summary: List content pieces
      description: Retrieve content pieces with optional filtering
      operationId: getContent
      tags:
        - Content
      parameters:
        - name: campaignId
          in: query
          description: Filter by campaign ID
          schema:
            type: integer
            format: int64
        - name: type
          in: query
          description: Filter by content type
          schema:
            $ref: '#/components/schemas/ContentType'
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: '#/components/schemas/ContentStatus'
        - name: language
          in: query
          description: Filter by language
          schema:
            type: string
            example: "en"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentPiece'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create content piece
      description: Create a new content piece within a campaign
      operationId: createContent
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPiece'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    ContentPiece:
      type: object
      properties:
        id:
          type: integer
          format: int64
        campaignId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/ContentType'
        originalContent:
          type: string
          nullable: true
          description: Original content text
        language:
          type: string
          description: Content language code
          example: "en"
        status:
          $ref: '#/components/schemas/ContentStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiGenerations:
          type: array
          items:
            $ref: '#/components/schemas/AIGeneration'
        translations:
          type: array
          items:
            $ref: '#/components/schemas/Translation'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    AIGeneration:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        generatedText:
          type: string
          description: AI-generated content
        promptUsed:
          type: string
          nullable: true
          description: Prompt used for generation
        modelVersion:
          type: string
          description: AI model version used
          example: "gpt-4"
        metadata:
          type: string
          nullable: true
          description: JSON metadata including analysis data
        createdAt:
          type: string
          format: date-time
    Translation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        targetLanguage:
          type: string
          description: Target language code
          example: "es"
        translatedText:
          type: string
          description: Translated content
        status:
          type: string
          enum: [draft, completed]
        qualityScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
          nullable: true
          description: Translation quality score
        createdAt:
          type: string
          format: date-time
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contentPieceId:
          type: integer
          format: int64
        language:
          type: string
          description: Language of the reviewed content
        status:
          type: string
          enum: [approved, rejected, needs_changes]
        feedback:
          type: string
          nullable: true
          description: Reviewer feedback
        reviewerName:
          type: string
          nullable: true
          description: Name of the reviewer
        createdAt:
          type: string
          format: date-time
    ContentType:
      type: string
      enum: [headline, description, blog_post, social_media, email]
      description: Type of content piece
    ContentStatus:
      type: string
      enum: [draft, ai_generated, under_review, approved, rejected]
      description: Status of content piece
    
    # Request schemas
...
    CreateContentRequest:
      type: object
      required:
        - campaignId
        - type
      properties:
        campaignId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/ContentType'
        originalContent:
          type: string
          maxLength: 10000
        language:
          type: string
          default: "en"
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request data"
            code: "BAD_REQUEST"
            timestamp: "2024-01-01T12:00:00Z"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-01T12:00:00Z"
...